fix(dhl recepción): castear $16/$20 a numeric en insert (inconsistent types deduced)
$16 (import_cost_mxn) y $20 (import_tax_mxn) se usaban en COALESCE($16,0) donde
el literal 0 los inferían como integer, chocando con su uso como columna numeric
→ "inconsistent types deduced for parameter $16". Se castea a ::numeric.

+++ entregax-backend-api/src/dhlController.ts
@@ -823,7 +823,7 @@ export const receiveDhlPackage = async (req: Request, res: Response) => {
         cost_payment_status,
         total_cost_mxn,
         status
-      ) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, NOW(), $14, $15, $16, $20, $17, $18, NOW(), $19, 'pending', (COALESCE($16,0) + COALESCE($20,0)), 'received_mty')
+      ) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, NOW(), $14, $15, $16, $20, $17, $18, NOW(), $19, 'pending', (COALESCE($16::numeric,0) + COALESCE($20::numeric,0)), 'received_mty')
       RETURNING *
     `, [
       inbound_tracking, secondary_tracking || null,
